+++ bluefin/CHANGELOG.md
@@ -1,3 +1,5 @@
+* Use `INLINE [0]` on `runDslBuilderEff` for improved performance
+
 # 0.10.0.0
 
 * Make `Ask`, `AskCapability`, `Await`, `JumpTo`, `Modify`, `Request`,